update, delete section iskelta i atskirus failus, pradeta su modalais kazkas daryht
